Pre-processing/requirements.txt
pandas
requests
google-cloud-storage
